Rename run batch simulations file
